Make sure both callback and promise Connection have serverProperties field.  Add test.
